HP-2419 Created PHPUnit test for BillingRegistryService
